backend/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "scripts*"]

[project]
name = "oneshot-fantasy"
version = "0.1.0"
description = "Python worker: stat ingestion + batch scoring (Symfony owns HTTP API)"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.0",
    # NBA ingestion
    "nba_api>=1.4.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
